Format genotype for BayEnv and BayPass

This function generates necessary genotype count formats for BayEnv and BayPass with a subset of SNPs

Usage

gt.format(gt, info, format = c("benv", "bpass"), snp.subset = NULL)

Arguments

gt

multi-vector. an imported data.frame of genotypes or genotype data frame generated by hetTgen or path to GT.FORMAT file generated from VCFTools

info

a data frame containing sample and population information. It must have “sample” and “population” columns

format

character. output format i.e., for BayPass or BayEnv

snp.subset

numerical. number of randomly selected subsets of SNPs. default = NULL subset

Value

Returns a list with formatted genotype data: $bayenv - snps in horizontal format - for BayEnv (two lines per snp); $baypass - vertical format - for BayPass (two column per snp); $sub.bp - subsets snps for BayPass $sub.be - subsets of snps for BayEnv

Author

Piyal Karunarathne

Examples

if (FALSE) vcf.file.path <- paste0(path.package("rCNV"), "/example.raw.vcf.gz")
vcf <- readVCF(vcf.file.path=vcf.file.path)
het.table<-hetTgen(vcf,"GT")
#> generating table
#>   |                                                          |                                                  |   0%  |                                                          |=====                                             |  10%  |                                                          |                                                  |   0%  |                                                          |==========                                        |  20%  |                                                          |                                                  |   0%  |                                                          |===============                                   |  30%  |                                                          |                                                  |   0%  |                                                          |====================                              |  40%  |                                                          |                                                  |   0%  |                                                          |=========================                         |  50%  |                                                          |                                                  |   0%  |                                                          |==============================                    |  60%  |                                                          |                                                  |   0%  |                                                          |===================================               |  70%  |                                                          |                                                  |   0%  |                                                          |========================================          |  80%  |                                                          |                                                  |   0%  |                                                          |=============================================     |  90%  |                                                          |                                                  |   0%  |                                                          |==================================================| 100%
info<-unique(substr(colnames(het.table)[-c(1:3)],1,8))
GT<-gt.format(het.table,info)
#> Formating BayEnv
#>   |                                                          |                                                  |   0%  |                                                          |=====                                             |  10%  |                                                          |==========                                        |  20%  |                                                          |===============                                   |  30%  |                                                          |====================                              |  40%  |                                                          |=========================                         |  50%  |                                                          |==============================                    |  60%  |                                                          |===================================               |  70%  |                                                          |========================================          |  80%  |                                                          |=============================================     |  90%  |                                                          |==================================================| 100%
#> Warning: non-unique values when setting 'row.names': ‘MA_10.24959~1’, ‘MA_10.26811~2’, ‘MA_10.26860~1’, ‘MA_10.26874~2’, ‘MA_100058.3823~2’, ‘MA_100058.4182~1’, ‘MA_100058.4582~2’, ‘MA_100058.6038~1’, ‘MA_100058.6049~2’, ‘MA_100058.7419~1’, ‘MA_100087.10845~2’, ‘MA_100087.11339~1’, ‘MA_100211.2284~2’, ‘MA_100349.38271~1’, ‘MA_100349.39403~2’, ‘MA_100349.39826~1’, ‘MA_100349.40809~2’, ‘MA_100472.21035~1’, ‘MA_100472.23448~2’, ‘MA_10047260.3511~1’, ‘MA_100474.3489~1’, ‘MA_10048.31196~1’, ‘MA_10049501.1836~2’, ‘MA_10049501.1986~1’, ‘MA_100525.12592~2’, ‘MA_10054999.1462~2’, ‘MA_10055697.1604~1’, ‘MA_100568.14098~1’, ‘MA_100568.15073~2’, ‘MA_100568.15133~1’, ‘MA_100606.48171~2’, ‘MA_100608.19198~1’, ‘MA_100608.19952~2’, ‘MA_10061551.352~2’, ‘MA_100687.19884~1’, ‘MA_100687.20041~2’, ‘MA_100687.20048~1’, ‘MA_100687.20166~2’, ‘MA_100687.20179~1’, ‘MA_10074020.1109~1’, ‘MA_10074020.1508~2’, ‘MA_10074020.1931~1’, ‘MA_10074020.3478~2’, ‘MA_10074020.3620~1’, ‘MA_10076142.5530~2’, ‘MA_10076711.1209~1’, ‘MA_10076711.1301~2’, ‘MA_10080382.2279~1’, ‘MA_10084120.1639~2’, ‘MA_10085137.5243~1’, ‘MA_10085137.6166~2’, ‘MA_10085137.6855~1’, ‘MA_10085137.6999~2’, ‘MA_100863.10544~2’, ‘MA_100863.10587~1’, ‘MA_100863.10655~2’, ‘MA_100863.10656~1’, ‘MA_100863.11117~2’, ‘MA_10087018.2764~1’, ‘MA_10090427.3278~2’, ‘MA_10090427.4475~1’, ‘MA_10090927.415~2’, ‘MA_10093414.463~1’, ‘MA_10096503.2131~2’, ‘MA_10096503.2534~1’, ‘MA_10096503.3170~2’, ‘MA_10096610.325~1’, ‘MA_10098066.21~2’, ‘MA_10098066.300~1’, ‘MA_10101.34237~2’, ‘MA_10101.34867~1’, ‘MA_101019.46121~1’, ‘MA_101019.46361~2’, ‘MA_101019.46661~1’, ‘MA_101038.26055~2’, ‘MA_101038.26786~1’, ‘MA_101038.27168~2’, ‘MA_101038.27364~1’, ‘MA_101038.31032~2’, ‘MA_101050.7265~1’, ‘MA_101050.7278~2’, ‘MA_10105635.2671~2’, ‘MA_101107.7846~1’, ‘MA_101113.10253~2’, ‘MA_101113.11204~1’, ‘MA_10111655.3831~1’, ‘MA_101150.11038~2’, ‘MA_101150.6693~2’, ‘MA_101150.6821~1’, ‘MA_10115334.2655~2’, ‘MA_101248.11533~1’, ‘MA_101248.11584~2’, ‘MA_101248.11587~1’, ‘MA_101248.12590~2’, ‘MA_10125264.222~1’, ‘MA_10126.10422~2’, ‘MA_10126.11785~1’, ‘MA_10126.12824~2’, ‘MA_10129927.2594~2’, ‘MA_10130047.3032~1’, ‘MA_101359.5955~1’, ‘MA_101359.6032~2’, ‘MA_101359.6245~1’, ‘MA_101359.6253~2’, ‘MA_101371.20121~1’, ‘MA_101371.22419~2’, ‘MA_101371.26631~1’, ‘MA_10137748.1513~2’, ‘MA_10144145.1482~1’, ‘MA_10146757.2274~2’, ‘MA_10146757.344~2’, ‘MA_10146757.846~1’, ‘MA_101553.10397~2’, ‘MA_101553.15758~1’, ‘MA_101553.15830~2’, ‘MA_101553.15934~1’, ‘MA_101553.16752~2’, ‘MA_10156303.803~1’, ‘MA_10156303.874~2’, ‘MA_10156761.2921~1’, ‘MA_101571.34150~1’, ‘MA_101571.34447~2’, ‘MA_10159942.165~2’, ‘MA_10161.11907~1’, ‘MA_10161.32075~2’, ‘MA_10161.34653~1’, ‘MA_10161.34871~2’, ‘MA_10161479.400~1’, ‘MA_10161479.502~2’, ‘MA_10161479.683~1’, ‘MA_10162666.3888~2’, ‘MA_10162666.4211~1’, ‘MA_10162666.5441~2’, ‘MA_10162666.5819~1’, ‘MA_10162666.6336~2’, ‘MA_10163169.336~1’, ‘MA_10165055.5538~2’, ‘MA_10165689.2148~1’, ‘MA_101664.5848~1’, ‘MA_101664.6142~2’, ‘MA_10166594.2025~2’, ‘MA_101691.16954~1’, ‘MA_101720.39752~2’, ‘MA_101720.39889~1’, ‘MA_101720.40134~2’, ‘MA_101720.43718~1’, ‘MA_101720.43931~2’, ‘MA_101720.44212~1’, ‘MA_101720.44338~2’, ‘MA_101720.44359~1’, ‘MA_10173312.1399~1’, ‘MA_10173312.1775~2’, ‘MA_10173312.2327~1’, ‘MA_10173312.2426~2’, ‘MA_10174633.3877~1’, ‘MA_101776.21787~2’, ‘MA_101803.15232~2’, ‘MA_101803.20433~1’, ‘MA_101803.21961~2’, ‘MA_101803.4037~1’, ‘MA_10182861.3140~2’, ‘MA_10182861.3156~1’, ‘MA_10187685.644~2’, ‘MA_10189777.135~1’, ‘MA_10190246.1449~2’, ‘MA_101915.12822~1’, ‘MA_101915.16048~2’, ‘MA_101915.16099~1’, ‘MA_101915.16292~2’, ‘MA_101915.18060~1’, ‘MA_101915.18304~2’, ‘MA_10194.13624~1’, ‘MA_10195410.1687~1’, ‘MA_10195410.1699~2’, ‘MA_10195410.3387~1’, ‘MA_10195746.2641~2’, ‘MA_10195746.2685~1’, ‘MA_101958.18348~1’, ‘MA_101958.22667~2’, ‘MA_101964.13105~1’, ‘MA_101964.14374~2’, ‘MA_101964.17546~1’, ‘MA_101964.17602~2’, ‘MA_101964.18933~1’, ‘MA_101964.19029~2’, ‘MA_10197194.3823~2’, ‘MA_10197194.4597~1’, ‘MA_10197194.5623~2’, ‘MA_10197194.6040~1’, ‘MA_10197194.7063~2’, ‘MA_101987.4615~1’, ‘MA_101987.5275~2’, ‘MA_101987.5657~1’, ‘MA_10199816.1289~1’, ‘MA_10201285.1409~2’, ‘MA_10201285.2173~1’, ‘MA_10206082.7672~2’, ‘MA_10206082.7762~1’, ‘MA_10206082.8155~2’, ‘MA_10206082.8500~1’, ‘MA_10206478.3496~2’, ‘MA_10206478.3504~1’, ‘MA_10206478.3729~2’, ‘MA_10206478.3820~1’, ‘MA_10207543.770~2’, ‘MA_10208000.3039~1’, ‘MA_10208488.1001~2’, ‘MA_102091.35337~2’, ‘MA_102091.36479~1’, ‘MA_102091.36867~2’, ‘MA_10211905.1189~1’, ‘MA_102148.52836~1’, ‘MA_102148.53566~2’, ‘MA_102148.53799~1’, ‘MA_10219080.1752~2’, ‘MA_10219080.2060~1’, ‘MA_102199.17408~2’, ‘MA_102199.19163~1’, ‘MA_102199.19473~2’, ‘MA_102199.20352~1’, ‘MA_102199.24258~2’, ‘MA_102199.24304~1’, ‘MA_1022588.909~1’, ‘MA_102307.31232~2’, ‘MA_102334.24075~1’, ‘MA_102334.24169~2’, ‘MA_10235.15900~2’, ‘MA_10235.16003~1’, ‘MA_102538.32570~1’, ‘MA_102538.32764~2’, ‘MA_102538.32824~1’, ‘MA_102648.5077~2’, ‘MA_102648.5698~1’, ‘MA_102682.27603~2’, ‘MA_102682.27921~1’, ‘MA_102682.28095~2’, ‘MA_102682.28300~1’, ‘MA_102702.17852~2’, ‘MA_102702.18191~1’, ‘MA_102702.18857~2’, ‘MA_102702.19343~1’, ‘MA_10270322.3967~2’, ‘MA_102714.21228~2’, ‘MA_102722.10095~1’, ‘MA_102722.10118~2’, ‘MA_10272340.1947~1’, ‘MA_10274.26875~2’, ‘MA_10274.27222~1’, ‘MA_10274407.2676~2’, ‘MA_10274407.2736~1’, ‘MA_10277343.4737~2’, ‘MA_10277343.6526~1’, ‘MA_102777.39050~1’, ‘MA_102777.39895~2’, ‘MA_102777.40040~1’, ‘MA_102777.40290~2’, ‘MA_102777.40376~1’, ‘MA_10277984.3463~2’, ‘MA_102800.6694~2’, ‘MA_102800.7081~1’, ‘MA_102800.7114~2’, ‘MA_10281009.1607~1’, ‘MA_102821.15350~1’, ‘MA_10285571.3955~2’, ‘MA_10285571.5455~1’, ‘MA_10285571.5804~2’, ‘MA_10285593.1459~1’, ‘MA_10285593.1477~2’, ‘MA_10285593.2717~1’, ‘MA_10285593.4011~2’, ‘MA_10285593.4577~1’, ‘MA_10293408.6140~2’, ‘MA_10293443.624~1’, ‘MA_10293443.740~2’, ‘MA_10296064.2939~1’, ‘MA_10297658.652~2’, ‘MA_10301686.4273~1’, ‘MA_10302565.251~2’, ‘MA_103063.16842~2’, ‘MA_10306333.3384~1’, ‘MA_10306333.3482~2’, ‘MA_10306333.3643~1’, ‘MA_10306333.3682~2’, ‘MA_10308515.545~1’, ‘MA_10310.17470~2’, ‘MA_10310.18008~1’, ‘MA_10310740.1171~2’, ‘MA_10312335.1227~1’, ‘MA_10313972.197~2’, ‘MA_103143.23705~1’, ‘MA_103143.25817~2’, ‘MA_103143.40031~1’, ‘MA_103156.38042~2’, ‘MA_10317671.3474~2’, ‘MA_10317671.419~1’, ‘MA_10317671.891~2’, ‘MA_10317671.964~1’, ‘MA_10319578.1803~1’, ‘MA_10320745.5600~2’, ‘MA_10320745.5860~1’, ‘MA_10322161.3413~2’, ‘MA_103256.9653~1’, ‘MA_103256.9752~2’, ‘MA_10326151.587~1’, ‘MA_10327183.206~2’, ‘MA_10329326.6585~1’, ‘MA_103310.26003~1’, ‘MA_10333127.2683~2’, ‘MA_10333529.3291~1’, ‘MA_10335992.1840~2’, ‘MA_10337341.2991~1’, ‘MA_103401.14082~2’, ‘MA_103401.14512~1’, ‘MA_103401.14539~2’, ‘MA_103401.28241~1’, ‘MA_10340850.980~2’, ‘MA_10343960.172~1’, ‘MA_10344908.26028~2’, ‘MA_10344908.26046~1’, ‘MA_103463.12884~2’, ‘MA_103463.15126~1’, ‘MA_103463.15194~2’, ‘MA_103463.15361~1’, ‘MA_10349496.597~2
#> Error in `.rowNamesDF<-`(x, value = value): duplicate 'row.names' are not allowed