Determine pairwise relatedness

Relatedness is determined according to genome-wide relationship assessment of Yang et al. 2010 equation 6, on a per sample basis (with itself and others), using SNPs.

Usage

relatedness(
  vcf,
  plot = TRUE,
  threshold = 0.5,
  verbose = TRUE,
  parallel = FALSE
)

Arguments

vcf

an imported vcf file in a list using readVCF or a data frame of genotypes generated using hetTgen

plot

logical. Whether to plot relatedness of samples against themselves, among themselves and outliers

threshold

numerical. A value indicating to filter the individuals of relatedness among themselves. Default 0.5 (siblings)

verbose

logical. Show progress.

parallel

logical. Parallelize the process

Value

A data frame of individuals and relatedness score \(A_{jk}\)

Details

According to Yang et al. (2010), out breeding non-related pairs should have a relatedness value of zero while the individual with itself will have a relatedness value of one. Relatedness value of ~0.5 indicates siblings.

References

Yang, J., Benyamin, B., McEvoy, B. et al. Common SNPs explain a large proportion of the heritability for human height. Nat Genet 42, 565569 (2010).

Author

Piyal Karunarathne, Klaus Schliep

Examples

if (FALSE) vcf.file.path <- paste0(path.package("rCNV"), "/example.raw.vcf.gz")
vcf <- readVCF(vcf.file.path=vcf.file.path)
#> Error: object 'vcf.file.path' not found
relate<-relatedness(vcf) # \dontrun{}
#> Error: object 'vcf' not found